From Christensen, ReedX K
  Avoided use of modulo (%) operator by modifying Rand() function.
Turning on random accesses in Iometer/dynamo was causing floating point
exception faults very reliably.  Compiled dynamo with -g and set 'ulimit -c
   unlimited' in order to get core file, and got some info from gdb:

   (gdb) where
   #0  0x00c8540b in __umoddi3 () from /lib/libgcc_s.so.1
   #1  0x08050f0a in TargetDisk::Seek ()
   #2  0x080501fe in Grunt::Do_IOs (this=0xb7f90008) at IOGrunt.cpp:1251
   #3  0x08050534 in Grunt_Thread_Wrapper (grunt=0xb7f90008) at
IOGrunt.cpp:992
   #4  0x00b3eb80 in start_thread () from /lib/libpthread.so.0
   #5  0x009c0dee in clone () from /lib/libc.so.6

   I believe the __umoddi3() call is the "%" or modulo operator inside
of this call function:
         ((TargetDisk *) targets[target_id])->Seek( access_spec.Random(
                 access_percent,
                    (unsigned int)targets[target_id]->Rand() % 100 ),
size,
                 user_alignment, user_align_mask );

  Found a mail thread on a postsql site that pointed out the tricky
numeric nature of modulo operator:
     http://archives.postgresql.org/pgsql-general/2005-05/msg01187.php

And this link:
     www.nabble.com/Re:-Remainder-(-)-operator-and-GCC-p1412329.html

From Ming Zhag

I did some changes base on Reed's patch. It is still unclear that if using
fmodl is a good idea. Though it can eliminate the exception.

Also it is not ok for other OS. Windows+VC can not compile successfully with it.